Pull Request 81 Dynamic Map Selection
Maps are dynamically selected by reading the map name text.
OCR function simplified and improved for map name detection.
Map names are no longer needed to be mapped to page and index
in static.py.  They only need to be added to setup.json.  Spaces and
underscores are ignored.

Process is as follows:
- at each map selection page:
  - screenshot each of the six map name sections of the images
  - process the text from the image
  - use the fuzz library to get a score compared to the map name setting
- go back to the page and index with the highest fuzz score and select it
  - if below a setting threshold throw an exception

Debug available:
- as with the round text detection, screenshots are saved off
- each map index is compared against a list of all map names to verify
it is detecting correctly